Added File Kubernetes_objects_function.py
